Refactor : Creating addressbook table in addressbook_service_db database

+++ SQL_Queries/AddressBookSystemSqlQueries.sql
@@ -1,6 +1,8 @@
+# UC1
 create database addressbook_service_db;
 show databases;
 
+# UC2
 use addressbook_service_db;
 create table AddressBook (
      FirstName varchar(25) not null,
@@ -9,7 +11,7 @@ create table AddressBook (
      City varchar(25) not null,
      State varchar(25) not null,
      Zip int not null,
-     PhoneNumber int not null,
+     PhoneNumber bigint not null,
      Email varchar(50) not null
      );
 show tables;
